sidenavlist test with no underscore and created sidenavlistitem tests
